# presentation/swagger/swagger.yaml
definitions:
  model.Ability:
    properties:
      alias:
        type: string
      isSkillArmorPenalty:
        type: boolean
      name:
        type: string
      shortName:
        type: string
    type: object
...
  model.Armor:
    properties:
      alias:
        type: string
      arcaneSpellFailureChance:
        type: integer
      armorBonus:
        type: integer
      armorCheckPenalty:
        type: integer
      book:
        $ref: '#/definitions/model.Books'
      cost:
        type: number
      description:
        type: string
      engName:
        type: string
      maxDexBonus:
        type: integer
      name:
        type: string
      speed20:
        type: integer
      speed30:
        type: integer
      type:
        $ref: '#/definitions/model.NameAlias'
      weight:
        type: number
    type: object
...
  model.Books:
    properties:
      abbreviation:
        type: string
      alias:
        type: string
      name:
        type: string
      order:
        type: integer
    type: object
...
  model.ClassForSkill:
    properties:
      alias:
        type: string
      book:
        $ref: '#/definitions/model.Books'
      isClassSkill:
        type: boolean
      name:
        type: string
      shortName:
        type: string
    type: object
...
  model.Domain:
    properties:
      alias:
        type: string
      book:
        $ref: '#/definitions/model.Books'
      childs:
        items:
          $ref: '#/definitions/model.Domain'
        type: array
      description:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  model.DomainInfo:
    properties:
      book:
        $ref: '#/definitions/model.Books'
      childs:
        items:
          $ref: '#/definitions/model.NameAlias'
        type: array
      description:
        type: string
      gods:
        items:
          type: string
        type: array
      innerDescription:
        type: string
      name:
        type: string
      parents:
        items:
          $ref: '#/definitions/model.DomainParent'
        type: array
      power0Description:
        type: string
      power0Name:
        type: string
      power1Description:
        type: string
      power1Name:
        type: string
      power2Description:
        type: string
      power2Name:
        type: string
      spells:
        items:
          $ref: '#/definitions/model.DomainSpell'
        type: array
    type: object
  model.DomainParent:
    properties:
      alias:
        type: string
      name:
        type: string
      power1Description:
        type: string
      power1Name:
        type: string
      power2Description:
        type: string
      power2Name:
        type: string
      spells:
        items:
          $ref: '#/definitions/model.DomainSpell'
        type: array
    type: object
  model.DomainSpell:
    properties:
      alias:
        type: string
      comment:
        type: string
      level:
        type: integer
      name:
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      message:
        type: string
      title:
        type: string
    type: object
...
  model.GetSkillsResponse:
    properties:
      skillsPerLvl:
        items:
          $ref: '#/definitions/model.SkillsPerLvlInfo'
        type: array
      skillsWithClasses:
        items:
          $ref: '#/definitions/model.SkillWithClasses'
        type: array
    type: object
...
  model.Helper:
    properties:
      alias:
        type: string
      isMain:
        type: boolean
      name:
        type: string
    type: object
...
  model.MagicItemForList:
    properties:
      alias:
        type: string
      book:
        $ref: '#/definitions/model.Books'
      engName:
        type: string
      name:
        type: string
      slot:
        $ref: '#/definitions/model.NameAlias'
      type:
        $ref: '#/definitions/model.NameAlias'
    type: object
  model.MagicItemInfo:
    properties:
      alias:
        type: string
      aura:
        type: string
      book:
        $ref: '#/definitions/model.Books'
      cl:
        type: integer
      constructionCost:
        type: number
      constructionRequirements:
        type: string
      creationMagicItems:
        type: string
      description:
        type: string
      destruction:
        type: string
      engName:
        type: string
      helpers:
        items:
          $ref: '#/definitions/model.Helper'
        type: array
      name:
        type: string
      price:
        type: number
      priceComment:
        type: string
      slot:
        $ref: '#/definitions/model.NameAlias'
      slotComment:
        type: string
      statistics:
        type: string
      type:
        $ref: '#/definitions/model.NameAlias'
      weight:
        type: number
      weightComment:
        type: string
    type: object
...
  model.NameAlias:
    properties:
      alias:
        type: string
      name:
        type: string
    type: object
...
  model.NpcInfo:
    properties:
      alignment:
        type: string
      book:
        $ref: '#/definitions/model.Books'
      features:
        type: string
      helpers:
        items:
          $ref: '#/definitions/model.Helper'
        type: array
      hitDie:
        type: integer
      name:
        type: string
      skillRanksPerLevel:
        type: integer
      skills:
        items:
          $ref: '#/definitions/model.Skill'
        type: array
      tableFeatures:
        type: string
      tableSpellCount:
        type: string
    type: object
...
  model.Skill:
    properties:
      ability:
        $ref: '#/definitions/model.Ability'
      alias:
        type: string
      name:
        type: string
    type: object
  model.SkillInfo:
    properties:
      ability:
        $ref: '#/definitions/model.Ability'
      description:
        type: string
      fullDescription:
        type: string
    type: object
  model.SkillWithClasses:
    properties:
      alias:
        type: string
      classes:
        items:
          $ref: '#/definitions/model.ClassForSkill'
        type: array
      name:
        type: string
      prestigeClasses:
        items:
          $ref: '#/definitions/model.ClassForSkill'
        type: array
    type: object
  model.SkillsPerLvlInfo:
    properties:
      alias:
        type: string
      isPrestige:
        type: boolean
      name:
        type: string
      skillRanksPerLevel:
        type: integer
    type: object
...
  model.Trait:
    properties:
      alias:
        type: string
      benefit:
        type: string
      book:
        $ref: '#/definitions/model.Books'
      engName:
        type: string
      name:
        type: string
      prerequisites:
        type: string
      type:
        $ref: '#/definitions/model.TraitType'
    type: object
  model.TraitType:
    properties:
      alias:
        type: string
      description:
        type: string
      name:
        type: string
      parentType:
        $ref: '#/definitions/model.TraitType'
    type: object
...
info:
  contact: {}
  description: Сайт с информацией для настольной игры Pathfinder
  license:
    name: Чудинов Андрей Дмитриевич
  title: Pathfinder Family
  version: "20220823"
paths:
...
  /api/allMagicItems:
    get:
      description: Список волшебных предметов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MagicItemForList'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список волшебных предметов
      tags:
      - API
...
  /api/armors:
    get:
      description: Список доспехов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Armor'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список доспехов
      tags:
      - API
...
  /api/domainInfo:
    get:
      description: Информация по сфере/инквизиции
      parameters:
      - description: Alias сферы/инквизиции
        in: query
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DomainInfo'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Информация по сфере/инквизиции
      tags:
      - API
  /api/domains:
    get:
      description: Список Сфер/Инквизиций
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Domain'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список Сфер/Инквизиций
      tags:
      - API
...
  /api/magicItemAbilities:
    get:
      description: Список волшебных предметов
      parameters:
      - description: Тип волшебных предметов
        in: path
        name: types
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MagicItemInfo'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список волшебных предметов
      tags:
      - API
  /api/magicItemInfo:
    get:
      description: Информация по волшебному предмету
      parameters:
      - description: Alias волшебного предмета
        in: query
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MagicItemInfo'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Информация по волшебному предмету
      tags:
      - API
  /api/magicItems:
    get:
      description: Список волшебных предметов
      parameters:
      - description: Тип волшебных предметов
        in: path
        name: types
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MagicItemInfo'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список волшебных предметов
      tags:
      - API
...
  /api/npcInfo:
    get:
      description: Информация по персонажу ведущего
      parameters:
      - description: Alias персонажа ведущего
        in: query
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NpcInfo'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Информация по персонажу ведущего
      tags:
      - API
  /api/npcs:
    get:
      description: Список персонажей ведущего
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.NameAlias'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список персонажей ведущего
      tags:
      - API
...
  /api/skillInfo:
    get:
      description: Информация по навыку
      parameters:
      - description: Alias навыка
        in: query
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SkillInfo'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Информация по навыку
      tags:
      - API
  /api/skills:
    get:
      description: Список навыков
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetSkillsResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список навыков
      tags:
      - API
...
  /api/traits:
    get:
      description: Список штрихов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Trait'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список штрихов
      tags:
      - API
...
  /armors:
    get:
      description: Доспехи
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Доспехи
      tags:
      - Site
...
  /magicItems:
    get:
      description: Все волшебные предметы
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Все волшебные предметы
      tags:
      - Site
...
  /traits:
    get:
      description: Штрихи персонажей
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Штрихи персонажей
      tags:
      - Site
...
swagger: "2.0"